웹 소켓 연결해서 게임 자동으로 업데이트 추가

+++ src/main/java/com/ssinhwa/gameserver/redisserver/dto/WPosition.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class WPosition {
-    private Float x;
-    private Float y;
+    private double x;
+    private double y;
 }
